Feature: move EffectiveMembership to effective_membership.rs
